+++ api/services/nexus.py
@@ -4,7 +4,7 @@ Nexus service to expose business logic of interacting with Nexus
 
 from urllib.parse import urlparse
 import requests
-from api.exceptions import InvalidUrlParameterException, ResourceNotFoundException
+from api.exceptions import AuthorizationIssueException, InvalidUrlParameterException, ResourceNotFoundException
 
 
 def fetch_file_content(access_token: str, content_url: str = "") -> bytes:
@@ -32,4 +32,6 @@ def fetch_file_content(access_token: str, content_url: str = "") -> bytes:
         return response.content
     if response.status_code == 404:
         raise ResourceNotFoundException
+    if response.status_code == 401:
+        raise AuthorizationIssueException
     raise requests.exceptions.RequestException